fix(convex): rename gdl-parser to gdl_parser for deploy compatibility
Convex rejects module paths with hyphens — only alphanumeric, underscores, and periods are allowed.
